Potential fix for code scanning alert no. 18: Implicit narrowing conversion in compound assignment

+++ msal4j-persistence-extension/src/test/java/com/microsoft/aad/msal4jextensions/CacheLockTestBase.java
@@ -108,7 +108,7 @@ public class CacheLockTestBase {
 
         list.sort(Comparator.comparingLong(a -> a[0]));
 
-        int sum = 0;
+        long sum = 0L;
         Long[] prev = null;
         for (Long[] interval : list) {
             Assert.assertTrue(interval[0] <= interval[1]);
